renamed:    meshspy-ui/src/lib/markerIcons.js -> meshspy-ui/src/lib/markerIcons.jsx
